Make mpmath work in sage on top of sage.Integer

+++ mpmath/libelefun.py
@@ -345,7 +345,7 @@ def nthroot_fixed(y, n, prec, exp1):
     start = 50
     try:
         y1 = rshift(y, prec - n*start)
-        r = MP_BASE(y1**(1.0/n))
+        r = MP_BASE(int(y1**(1.0/n)))
     except OverflowError:
         y1 = from_int(y1, start)
         fn = from_int(n)
